pyproject.toml
[project]
name = "kn-rag-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.120.4",
    "gradio>=5.49.1",
    "langchain>=1.0.3",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=3.0.0",
    "langchain-text-splitters>=1.0.0",
    "markdown>=3.9",
    "nest-asyncio>=1.6.0",
    "pydantic>=2.11.10",
    "pypdf>=6.1.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "unstructured>=0.18.15",
]
